% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesisdown.R
\docType{package}
\name{tisemdown}
\alias{tisemdown}
\alias{tisemdown-package}
\title{tisemdown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
Using the bookdown package and unofficial LaTeX thesis templates from
    Tilburg University, this package aims to encourage reproducible research using
    simple Markdown syntax while embedding all of the R code to produce plots
    and analyses as well.
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_epub}{

Creates an R Markdown thesis template as an ebook
}

\author{
\strong{Maintainer}: Lachlan Deer \email{lachlan.deer@gmail.com}

}
